requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
requests>=2.32.0
joblib>=1.4.0
scikit-learn>=1.4.0
